AppDV rules for add/edit fields + normalize allow_add/allow_edit
The values for allow_edit and allow_add are now normalized and
populated according to existing rules (just for ease in testing
them). Additionally, AppDV now properly considers their values before
starting a field edit, which can be different for edit vs add and
behave properly.

Completing the rules for hiding UI elements is nearly done but not
quite. But at least now fields don't show up when they are not
supposed to.

Still in progress...