allow_view and no_column now normalize with allow_edit/allow_view
Added logic/consideration to pre-populate and normalize 'allow_view'
along with allow_edit and allow_add (changes started in 9e07f9fc21)

Now all 4 of the column attrs which effect eachother's defaults are
evaluated and populated together, in perl, on the backend:

 * no_column
 * allow_edit
 * allow_add
 * allow_view

The rules for how these interact are maintained (intended to be,
anyway) - see RapidApp::Manual::TableSpec - but are now colsolidated.

Unfactored several places higher in the stack which had this logic
inline, and is now redundant.

Unfortunately there are no unit tests to cover this domain (since it
would require browser/client JS testing which we don't have yet) so
it is possible for there to be bad edge cases, however, I did test
the combinations described in docs and many other places and
confirmed the behaviors remained the same

(set dev ver 1.0601_18)