AJAX EXCEPTION DIALOGS NOW CONSTRAINED TO ORIGIN PANEL + misc - #6
The special case of an ajax exception now checks to see if it can
identify the panel/cmp which generated the request, and if found,
uses smartRenderTo to constrain the popup to that panel/tab.

Also cleaned up and refactored the timeout case to not display both
a timeout dialog and a generic exception dialog which is redundant.
Now, only the timeout msg is shown, which has been spiffed up and
also includes the statusText (which in all cases I've ever seen
is always "transaction aborted")

Finally, made some tweaks to handle the case of a window with a
smartRenderTo loading when it is not the active tab. The window is
not able to size itself correctly in this case, and so it would
render totally off the screen (appearing as if it didn't render at
all). 95% of this problem has been solved by enabling the
'constrain' option which keeps it within the target element.
However, it's not perfect; the window appears in the top left
instead of center. But this is a much less urgent problem (however
I do still plan to revisit)

See also in-code comments.

(set dev ver 1.0601_21)