DEFAULT COL ORDER NOW PUTS SINGLE RELS NEXT TO THE FK - fixes #163
The default column order now places foreign-key single relationship
columns in the same spot (immediately after) their underlying, real
FK column (which is typically hidden anyway), instead of at the end
of all the physical columns. This is closer to the real, intended
order of the columns specified by the table schema, since FK-based
single-rels are the main representation of the FK column in the
interface.

(set dev ver 1.0601_22)