Generalization/centralization of column rendering (IN PROGRESS)
Generalized DataStore rendering + code-path to pass through a new,
single control point within the datastore plus plugin object.
Refactored the 3 main datastore-based components, Grid, AppDV and
PropertyGrid to fire through this new path.

This is currently passive (simple passthrough with no behavior
changes so far) but this provides a new powerful hook to be able to
extend column rendering models which will come in next commits. The
first planned feature to take advantage of this hook is "dirty"
rendering codepaths.

IN PROGRESS...