DIRTY (i.e. MODIFIED) DISPLAY OF SINGLE REL COLUMNS NOW HANDLED
This has been a very long-time want/limitation which has never been
solved until now due to the complexity of the problem. Basically, for
datastore grids/DVs which are configured with persist_immediately
turned off, changing the value of a single-rel column would not be able
to render the new value until being persisted because these columns
render via a different column (the display_column), joined through the
underlying FK, and this updated information is not available until
refresh from the server...

While there is no direct way to overcome this 100%, since the data
simply is not yet known to the client, the **editor** which is used to
modify the value has a pretty good idea of what this display value
should be, since for single-rels, this is used for the selection of the
new value in the first place. So, this commit now provides a mechanism
for the editor to **report** that value back to the datastore for it
to be able to use to render the value until the real data is available
after round-trip/save to the server.

So, while this value is not guaranteed to be correct, since only the
server knows that for sure, it is guaranteed to be better than what it
was -- the old value. Like other cases of marking dirty data inline (
like the little red triangles), temp values rendered like this are
marked by lightened italic text, with dotted red underline.

I had this idea a long time ago, but I couldn't implement it without a
major refactoring of the way column rendering works, which has now
been done (actually in the last commit, 82d4dee0ace) mainly to support
this feature.

This only works currently for update, freshly created rows are still
TBD (but that case was never as important because there is no old value
to incorrectly display)

(set dev ver 1.0601_23)