AppDV: permit native "submit" target element clicks through
Add another exception to the default Ext.DataView behavior of blocking
navigation element click events to allow all target clicks with
the type="submit" attribute (i.e. form submit buttons). We already
allow <a> clicks only if target="_self" however for the case of a
form submit there is no special internally handled nav case, so we
now just allow it through per default. We assume that if the dev user
puts a form with a submit in their template, they want it to work.

(set dev version 1.0601_26)