AppDV: Refactoring/fixes to click decicion logic (update to prev)
Refactored the structure of the click "decision" of onBeforeclick
which decides if teh click event needs to be handled by AppDV or
by the browser to support more complex logic (now via nested
recursive func call) and added extra support to handle the case of
a user using a <button> custom markup as a AppDV control -- for
that case, no, its not a submit and we do need to locally capture it
(see prev commit). Now it detects the appdv generated 'clickable'
class and always captures for that case. Additionally, the custom
target="_self" (or any declared value) override which applies to <a>
tags can now be used for any type of tag to indicate that it should
be handled by the browser and not locally by AppDV.

(set dev ver 1.0601_27)