updateCssForStoreAPI() now only runs if rendered (bugfix)
Fix regression introduced in 83c15d44f0ebf2c2c5ebbf1c4b1d33e2cab34f36
which calls the DataStore updateCssForStoreAPI() function on every store
metachange to NOT call the logic unless the component is rendered. There
are cases (likely race conditions) where the metachange event is called
before the render has completed, and things break when this happens.
Modified the logic in the function to return unless the component
rendered flag is true.

This was a bug discovered in a legacy application which involves a
slower page consisting of multiple and large datastore components
(AppDVs and grids) which likley triggered a race-condition

(set dev ver 1.1000_02)