Prevent deep recursion bug in reportDirtyDisplayVal feature
An interaction between the recently added reportDirtyDisplayVal feature
and the previously existing functionality for async/background task of
fetching the display value in DataStoreAppField-based combos (like the
grid selector editor for single rels) setup a deep recursion via the
getValue() call. To prevent this, a special flag is now set during the
async query lifecycle which is now checked to abort recalling the
display value lookup again.

This bug expressed itself in a legacy app in a specific section (when
loading an edit form) and bisected to
a2e78c447c525ef36a7df1e9fc66c788b797703c, however, that commit just
enabled the codepath, so it is likely the regression existed in a
previous commit in other areas

(set dev ver 1.1000_03)