Fix col summaries for multi-rels /w old, loadCnf inlineLink - #93
When support for Column Summaries on multi-rel columns was added, it
never worked for the case of multi-rels which used the fallback
"inlineLink" renderer because column summaries call the original columns
renderer, which threw an undef exception, because the "record" argument
is not supplied in that context. This has been fixed in this commit.

Note that this is a very uncommon case, most/default mullti-rel columns
use the newer Ext.ux.RapidApp.DbicRelRestRender which already handled
this case, but there ARE still scenarios which must fall-back to the old
style due to certain limitations which still exist.

(set dev ver 1.1000_04)