smartRenderTo now implies the parent tab - #6
The places which use the newly added 'smartRenderTo' to constrain to the
local component, now are constrained to the parent tab of that
componenent. This has been done for more reliable behavior, since the
'renderTo' option of the built-in Ext.Window which smartRenderTo enables
doesn't seem to be able to handle all the different scenarios which are
possible, so instead its just applying to the overall tab (if found).
This is still more specific than the original behavior which was the
whole browser.

See in-code comments for additional info

(set dev ver 1.1000_06)