AppDV: new opt 'multi_editable_autofield' for legacy
Starting in 1f1b2a720a5257fee8c5bbfdc376c8b29e54298a only the first
instance of an autofield column in the template is editable. However,
there are cases where you do want to specify the same column multiple
times and allow it to be editable multiple times. For this case, the new
bool attr 'multi_editable_autofield' can be turned on (defaults to
false)