Bootstraps both regen_schema.pl and regen_schema_from_ddl.pl
When a RapidDbic app is bootstrapped it is given the script
'devel/regen_schema.pl' which can be called to regenerate the schema
classes via the connect info in the model (also bootstrapped). If the
app was bootstrapped using the "--from-sqlite-ddl" option then the
app is also bootstrapped with 'devel/regen_schema_from_ddl.pl' which
is the same as regen_schema.pl except the SQLite db file is first
overwritten from the current copy of the ddl (in the sql/ dir). Since
this is a destructive process, the user is prompted to confirm before
proceeding.

(set dev ver 1.1004_03)