RapidDbic bootstrap improvements to generated model class
The Model class which is generated by the RapidDbic bootstrapper now has
the following improvements:

 * Uses RapidApp::Util::find_app_home instead of Catalyst::Utils::home
 * Loads the new Moo role RapidApp::Util::Role::ModelDBIC

This new role adds the method _one_off_connect() which can be used to
connect to the schema using the config contained within the model class.
This is useful for scripts to be able to use the model class directly
without the overhead of loading the whole app class, which can have a
slow startup. Now with the new RapidApp::Util::find_app_home() call to
find the home dir for the case of SQLite models, this just works.
Before, this technique didn't work because the path to the sqlt db file
couldn't be found when the app class wasn't loaded.

(set dev ver 1.1004_05)