Make loader default db_schema '%' for all - #152
At some point a default changed regarding Schema::Loader with Postgres
databases to only scan the 'public' schema. This can be changed with the
'db_schema' option and set to '%' for all schemas which has now been
done in the RapidApp (MetaKeys) customer loader class. For rdbic.pl, the
schemas can be restricted using the already added --limit-schemas-re
option.

TODO: this needs to be formalized and added to rapidapp.pl also