Expose 'loader-option' in rapidapp.pl/rdbic.pl - #152
Exposed options in both rdbic.pl and rapidapp.pl:

  --loader-option
  --connect-option

These options were already available to rapidapp.pl, just undocumented.
Now the POD as been updated for both, as well as attributes with POD in
the rDbic object class.

The primary reason this is being added now is to provide access to the
'db_schema' option which is relevent to databases which have the extra
"schema" concept like Pg and MSSQL. Previously this need was attempted
to be solved by the '--limit-schemas-re' option (only in rdbic.pl),
however, this option is still dependent on an appropriate 'db_schema'
value which just can't be made to match "all" across all types of
databases. The default value of '%' works for Pg, but does not work for
MSSQL (it would require '{ "%" => "%" }'). For this reason, the '%'
default added yesterday has been removed.

The other reason is that when db_schema='%' is used in combination with
--limit-schemas-re, the scan still happens across all schemas and then
filtered out later, and this can waste time depending on how many other
schemas.

Now, to scan only the 'foo' schema, instead of using this rdbic.pl
option:

  --limit-schemas-re foo

You should instead use this option:

  --loader-option db_schema=foo

Which also works the same for the RapidDbic helper in rapidapp.pl.

For now the limit-schemas-re (and exclude) options are being left,
since they are more flexible, however, they  will probably be removed in
the future just for the sake of less complexity.