Fix/refactor ''/-as handling for MSSQL, etc - fixes #165
Fixed regression introduced in 78523655b8e which broke sorting on
virtual columns. Had to revisit this change which was done for MSSQL to
also consider the way virtual sorting is done. Stripping the -as seems
to break the ability to use the name in the sort (even though it is
still in the 'as' but that is for DBIC).

Now this is handled in a more broadly compatible manner, which is
described in more detail within in-code comments.

Tested sorting, searching on different kinds of virtual columns on all
of MSSQL, Pg and MySQL and all seem to be working correctly now as of
this commit.

(set dev ver 1.1006_01)