Refactor col summaries to use named keywords
Further refactored the column summary functions to be named
labels/keywords translated on the backend, but still cached in the
client. Now the client is not sending through raw SQL but a name string
which it looked up. This structure will prevent SQL injection (however,
custom functions are currently allowed by default which is by definition
SQL injection by design, however, it can now be tunred off and be safe,
once that is done)

Additionally, the functions are now translated according to the backend.
So far, only the MySQL versions of the date funcs have been implemented,
but at least now when this is ran on other backends it will display a
more useful "Unsupposed" message. The spots in the code now exist to add
for the other backens.