EXPERIMENTAL: add new 'CfgWriter' to bootstrap process
The new class/module 'RapidApp::Util::RapidDbic::CfgWriter' (see POD) is
now used to set the 'TableSpecs' configs in the database model when
bootstrapped via the 'RapidDbic' helper. This sets more verbose default
settings according to teh actual tables/columns present in the schema.

Additionally, new apps are now bootstrapped with the script:

  devel/model_db_updater.pl

This script replaces the old regen_schema* scripts and adds options to
regenerate the schema, database from DDL, and/or TableSpecs configs.
POD/usage has been added to guide the user-developer, but is still
preliminary.

This is still being worked on and tested and should be considered
EXPERIMENTAL!

In progress...

(Set dev version 1.1100_01)