Started 'rau.pl' script/system (umbrella utility script)
Added a new script 'rau.pl' as a general-purpose entrypoint for
command-line tools which will be added going forward. The idea is to not
pollute the user's system with anymore independent scripts, and instead
define a sub-command system (much like the way git works).

So far, the only module which has been added is 'modal-update' which
exposes RapidApp::Util::RapidDbic::CfgWriter on the command line. This
is reduntant for newly bootstrapped apps which now get created with
'devel/modal_db_updater.pl', but now the same functionality can be
accessed for existing apps using rau.pl.

See POD for additional details

(set dev version 1.1100_02)