FIX BUG PREVENTING COL PROFILES FROM SETTING allow_add/edit
Removed very old logic which sometimes toggled the values of allow_add
and allow_edit to true based on the colspec settings of the module. This
is now redundant with the normalization of allow_add/edit in the Column
class, but worse, this logic was getting called before column profiles
were applied, which effectively prevented column profiles from setting
allow_add or allow_edit to false.

This is a very old bug. So it is possible that older code out there is
relying on the bug, so this *might* be an API change for same cases
(note it is very unlikley that any users outside of the internal dev
team will be affected, but it is possible)

(set dev version 1.1100_11)