PROPERLY FIX BUG PREVENTING COL PROFILES FROM SETTING allow_add/edit
The commit
https://github.com/vanstyn/RapidApp/commit/2dccadc6f307e55c2f0cb9d1974bd0c68cbdc033
from May 30 was a regression which fixed the long-standing problem of
column profiles not working correctly when they set allow_add/allow_edit
to OFF, but created a new bug which incorrectly made columns not addable
when editable is off. This is the case of these settings being set via
the creatable_colspec and updatable_colspec in the local module, which
is an order of operations problem -- the normalization of allow_add/edit
doesn't happen until AFTER the module code modifies these settings
according to its local colspec settings.

The fix, which is ugly and temporary, is that the column profiles are
now considered in the module code itself. This is the best solution,
before a long-planned refactor of this stuff can be done.

See also in-code comments.

(set dev ver 1.2000_04)