Use func ::_table_name_safe() instead of method call in some places
Changed a few calls to the recently added _table_name_safe func/method
from method call to absolute package/function call to handle rare cases
when the component class is not loaded into the result class. This was
already done in some places for this reason before, but I received a
report of the call in DbicLnk causing an error for an end user/dev. This
turned out to be related to a bug in their code, however, it did show
that this scenario is possible, so, better safe than sorry. Only
downside of calling it like this is uglier code