** BEGIN NEW Template::Provider::Store API **
Started writing new API for user-defined, internal template storage.
This is being split out from the Provider class into a new layer called
the 'Store' class which is supplimental to the built-in, file-based
templates. This design is being pursued for the following primary
reasons:

 1. The Provider class is more complex within RapidApp than within TT,
    so it makes sense to give the user a more clean API

 2. The Store API doesn't replace the built-in provider, which is
     relied on for rendering RapidApp/system templates. The idea is the
     Store class can own a certain portion of the template
     path/namespace, while leaving the rest of it to be rendered the way
     it already is. This should provide a nice/clean way to address the
     need for user-defined template backends without mucking with the
     existing guts which rely on the template controller system

So far, all that has been done is defining of the methods and hooking
into a concept of asking the Store class if it owns any templates for
all the basic provider methods (create, update, fetch, etc). The dummy
base class implements all these methods as unimplemented (i.e. sub {...})
and claims to own no templates. The idea is that the user-defined class
will need to handle these methods.

This is very preliminary.

IN PROGRESS...

(set dev ver 1.2101_50)