Templates: Define new class-based 'post-processor' concept
The concept of "post-processing" of a template means handing stuff like
markdown. This was hard-coded in the Context class, but now it is
handled by a simple postprocessor class which is now determined by the
Access class on a per-template basis. The default logic still just
checks the file ext (with .md for markdown still the only built-in) but
the new API is designed to allow the access class to make more
sophisticated decisions

This is still experimental...