AppTree: begin work on node_types feature
Preliminary work to support the concept of multiple 'types' of nodes
which can be defined via new attr 'node_types'. The idea is to allow the
user to add and modify nodes of varied types, which can have additional
attributes besides 'name'

With this commit, the basic structure is in place which will change the
'Add' menu/button into a sub menu listing all the types. However, the
actions are not yet different and no special attributes are handled.

IN PROGRESS...