SavedState links use direct URL when there are no params/state_data
To support the new simple "link" mode, SavedState rows which have no
params or state_data now navigate directly to the saved url instead of
navigating to the url of the SavedState/view itself. These rows are just
links and point point anyplace. While this wasn't the orifginal intent,
those mode is still consistent and the behaviour can be considered
expected, since if there is no state data, there is no reason to
hide/alias the URL.

The Manage navtree when creating new 'link' nodes now creates SavedState
rows in this manner