Tree: Round-trip working modify custom attrs
For custom node types, the Rename routine now works to persist the
changes on the backend (note: requires its own implementation, which is
being done for the Manage NavTree in this commit). Also added support
for a 'new_attributes' in the rename return packet which will update the
nodes attributes. There was already one-off handling for the node text,
but now we need to be able to update other attributes which we now do.
Note that this functionality is just so that we can update the interface
in-place in a nicer way than calling a reload on the node, which we
could do, and is still a fall-back