** Manage Navtree can now choose icon from list of icon assets
Using new features just added to Catalyst::Controller::AutoAssets in
v0.40, a new, generated JavaScript asset is built which sets a global JS
variable containing a list of all the icon names that were generated by
the IconSet assets. And this global is now used in a new,
'ra-all-icon-assets-combo' combo/field which provides a dropdown with
all the icons to be able to select from. Finally, this new field is now
used in the Manage Navtree, so when users create/modify Custom Links,
they can not only set the icon, but be able to browse and choose from
one of the iconc which are available.

(set dev ver 1.2101_02)