'hidden' col profile explicitly sets allow_add/edit off
This is the bug that just won't die...

A regression caused by 10f05b76a25e5ff32686115a86d565bdc66861cd made it
so that the 'hidden' column profile no longer implied allow_add and
allow_edit to be false. However, this is actually expected based on the
change in that commit. It seems as though it was an oversight having not
explicitly set those options at the time. Hopefully its really finally
fixed now

(set dev ver 1.2111_01)