JSON MIXEDENCODER CODEREF SUPPORT (literal sql default_value now honored)
EXPERIMENTAL:

The RapidApp::JSON::MixedEncoder now supports encode-time evaluation by
inflating CodeRef values by calling them. This new feature is now being
used by the "default_value" which sets the editor (default)value for
columns. In the case of ScalarRef default_value, such as
\'current_timestamp', this means literal SQL. We now try and actually
ask the database for the current evaluation of the expression at the
time of encoding, which means at the time of the request (i.e. when the
add form is loaded). This allows the editor to be correctly
pre-populated with the current time, AS RETURNED FROM THE DATABASE.

See in-code comments for more info