Redesign MdEditor - now working w/o being so buggy!
The design of the SimpleMDE MdEditor has been totally refactored. Still
uses an iframe, but now the initialization happens within the iframe
itself via inline <script>, and this seems to have solved buggy errors
that were happen such as not being able to highlight text, and other
anomolies.

Left the iframeTextField which is no longer used as a working reference
implementation for iframe plain textarea editor fields.