MdEditor: upload handles images vs files + misc
The upload logic detects if the file is an image and if it is it uses
the image syntax. This is preliminary, the plan will be to do more than
this, including calling the alternate simplecas image upload and taking
action on the response which includes things like dimensions. Also will
consider inserting using <img> tags instead of markdown image syntax to
be able to control more aspects of rendering, namely max width...

Also cleaned up and simplified some of the code/markup and also chose a
better icon. Still WIP

IN PROGRESS...