Role::ModelDBIC injects schema->_ra_catalyst_origin_model
Because its been such a common need, I decided to go ahead and expand
the API to provide a consistent mechanism for DBIC schema/result classes
to access the origin catalyst model which composed them. This has been
done by adding a new method directly to the schema class, however, it
calls an item of the blessed/connected schema object, so this should be
safe/scoped and still work in case the same schema class is used
multiple times.

This was added in order to get the app class. This can be done already
via RapidApp->active_request_context, however, that only works in the
context of a request, and more and more times comes up (e.g. via backend
scripts) that the app class is needed and not always during a request.
So, this has finally been added.

(set dev ver 1.2111_56)