ra-mo-expandable-max-height: improvements + enabled for AppGrid
Lots of improvements to the ra-mo-expandable-max-height logic which does
a better/smarter job of handling and dealing with the 'max-height' style
property of the element in question (since max-height is how it works).
Also added special logic to apply/use a max-height value by looking for
a special coded class name. 'mh-200' means max-height: 200px, which is
now the hard-coded rule for the grid cells