MdEditor: fix some bugs with multiple editors breaking
Moved the "autoCreate" config from the class definition to be set on the
object during init. This fixes a bug where multiple editors ended up
sharing the same autoCreate which was getting globally modified with the
id of the first one, wich caused all sorts of problems.

Also added a cleanup function, although I'm not sure if it really is
needed (but it shouldn't hurt)