DataStorePlus: new option 'reload_on_show' + AutoPanel relayEvents
Added new datastore option 'reload_on_show' to reload teh store every
time the tab is (re)shown. Default is off. This is here as a poor man's
way to ensure a grid/dv is updated after one of its records has changed
from a related page. The long-term solution for this need is "tied
datastores" which would provide a way to connect store events from the
multi to page module (and, I've never added this option before because I
have been playing to add this feature). However, there are other times
when this option might be useful, and so now its available.

AutoPanel now relays events. In order for this to work properly, I also
had to add code for the AutoPanel to relay the 'show' event to its inner
content. This is something that should have been happening all along
probably, just never came up before