AuthCore: ?to= and to_referer support '_fragment'
Added mechanism for the client to be able to supply the 'fragment' (URI
nomenclature for #hash part of URL) to the ?to= redirect option as well
as the to_referer auto redirect logic.

For example, if the current url/page is '/some/page', linking to the
following URL (i.e. <a href=):

 /auth/to_referer?_fragment=add-comment

Would upon successful auth redirect the user back to:

 /some/page#add-comment

This is needed because the # fragment part of a URL is not sent in the
referer header.

Additionally the ?to= logic is now properly uri encoded/decoded so it
can now support its own query string values and work properly