AppDV: 'appdv-fill-absolute' now more fully supported
Wrapping an edit field in a tag with class 'appdv-fill-absolute'
triggers special logic tell AppDV to make the field fully fill this
outer wrapper. This provides a way to escape from all the crazy sizing
logic within plain HTML in an AppDV.

Earlier this was just CSS rules, but now the AppDV js code has been
updated to to correctly handle this case. Also, CSS has been updated to
handle the edit/save/cancel icons to correctly position them at the
bottom-right of the field. There is even an extra class that can be
added with appdv-fill-absolute - 'space-for-icons' - which will give an
extra 17px at the bottom for these icons/controls. Otherwise, they will
just overlap the field (but, now at least overlap in the bottom right,
prior to this change they were overlapping in the top-left which was
obviously a much bigger issue

(set dev ver 1.2111_63)