appdv-fill-absolute now works for normal (not 'big') fields
Also made API change, setting height:100% on the top element which has
the appdv-fill-absolute class. This is needed for the normal fields to
work when they do not have some external mechanism to give them a
height. The problem is when we do have a heigt via absolute position
with an offset, such as "top:40px" -- in that case, the height 100% will
cause the element to be 40px too big.

As a solution, when doing these offsets, it will be needed to set a
hgiht using calc(), for instance:

 height:calc(100% - 40px);

solves the problem and now it works in both cases

(set dev ver 1.2111_64)