MdEditor: show text with important toolbar icons
Added extra code/logic to apply text to the upload and side-by-side
preview buttons since these are important features that we want the user
to see without having to mouse-over. Also, the upload button text
declares that drag/drop is supported, which is actually the main thing
which I wanted to accomplish by this change.