Bugfix in dedicated-add-form/autoload_added_record post-save logic
I *think* this solves the problem where sometimes the new page doesn't
load before the add form closes itself, but I'm not 100% certain. There
may still be a race condition, but after this change, I cannot seem to
be able to get it to fail. So, moving on for now, if it happens again,
will revisit again

(note, this is still ugly no matter what, as is documented in the
comments)