ra-mo-expandable-max-height no longer default in AppGrid
Moved the logic which enabled cell max height out of the core AppGrid
and into a separate plugin 'ra-grid-expand-max-cell-height'. This logic
is now inactive unless this plugin is manually loaded in consumer grid
module... This was done because this can cause too big of a performnce
hit when there are lots of cells (rows x columns) on a single page. Will
revisit this later to come up with some kind of selective or more
efficient logic. In the meantime, the plugin is now available.