FORCE LEFT JOIN OVERRIDE MOVED TO OUTER SUB
In cab4a6732887df79d060a1204631a8df1cecd673 experimental code was added
to force all joins to left joins in DbicLnk (i.e. grid) modules. See
that commit for the rationales for this change, and see the comment
thread in GitHub for additional discussion.

This code has been working, however, the location it was originally
placed in ->rs_all only effected the rows, and not the associated count
and summary function calls, which could lead these to display the
incorrect result, excluding certain rows.

Moved the override to the outer sub, ->read_records, so it is in effect
for all three potential queries (rows, total count and summaries) for
each page load

(set dev ver 1.3003_01)