Use correlate in proper DBIx::Class::Helpers package
The correlate functionality was originally copied out of code from
DBIx::Class::Helpers because it was only implemented as a ResultSet
component at the time. Since then, @frioux as released a new version
with an option for a simple function interface (thanks fREW) so we're
now calling back to this instead of maintaining our own code
implementation.

Note: the latest version of helpers correlate also pre-supports the next
DBIC release which futher deprecates (and breaks in some cases)
_resolve_condition(). A similar change will be needed in RapidApp which
still relies on calls to ->_resolve_condition