SWITCH FROM _resolve_condition TO resolve_relationship_condition (if recent dbic)
The next version of DBIC - not yet released as of this commit - provides
a new official public method ->resolve_relationship_condition which is
better than the legacy ->_resolve_condition, and, ->_resolve_condition no
longer works the way it used to, and it no longer works in the one place
we were still using it, in multi-rels with a 'where' in attrs (which is
still a known limitation in DBIC, even as of the most recent devrel).

For this one special case, we now conditionally use the new
->resolve_relationship_condition method *if* it is available, and if it
is not, it means its an earlier version of DBIC and still calls
->_resolve_condition as before. Note - this is the exact same approach
to conditionally add this support as was done in correlate() in
DBIx::Class::Helpers, which is what this was based on.

Also replaced the localized __WARN__ override in the block that still
calls ->_resolve_condition to target the specific warning we want to
mute using the sigwarn_silencer() in DBIx::Class::_Util (thx @ribasushi)

(bumped dev version to 1.3004_01)