NavCore org tree can now change icons for all types + refactoring WIP
The organize navtree section now has support for changing the icon of
any objects in the tree, which now come in 3 varieties -- folder,
search, link. The 'link' type was the most recent to have been added and
it was the only one which could be customized, but now all three can.

This also involved some cleanup and re-organization of code.

Still have a number of outstanding tweaks to be fixed.

IN PROGRESS

(set dev version 1.3005_02)