NavCore: Add IMPORT button for 'default_public_nav_items' to Organize Tree
Exposed the routine which imports the 'default_public_nav_items', if
configured, to be runnable on-demand by admins from the Organize Tree.
By default, on app start up, this import runs automatically ONLY when
there are not any existing items in the public tree. But since it is
useful to be able to "get back" this default set of menu items without
having to first delete everything and restart the app, the nteractive
feature has been added.

(set dev version 1.3005_04)