AppTab: loadContent better support for simple urls (fallback to hashnav)
Added code to the AppTab loadContent() (the handler for loadContentCnf)
to perform a simple hashnav instead of the full-blown loadTab code when
it isn't needed, which it isent when there is a url with no params in
the autoLoad, and the AppTab is the main-load-target. This fixes loading
of tabs/urls via a loadContentCnf when the url contains a query-string,
and just behaves in a more reliable manner. This specifically came up
with the new Custom Link support added to the NavTree

See also in-code comments