RapidDbic now uses default loader-option preserve_case=1 (instead of 0)
The default for Schema::Loader is preserve_case=0, however, it really
makes more sense for RapidApp to use preserve_case=1 as the default so
that column names match whatever the actual database is, which may very
well use mixed-cased names, and is more likely to be the expected
behavior (especially for rdbic.pl). There are good reasons to use
lowercased columns, however, it makes less sense for that to be the
default.

The old behavior can still be accessed by supplying the following to the
command-line of rdbic.pl or rapidapp.pl (when using RapidDbic helper):

  --loader-option preserve_case=0

Updated the rapidapp.pl usage examples to show the use of
preserve_case=0 in one of the examples

(set dev version 1.3100_01)