AuthCore: login passthrough redirect converted to pure JS
The built in login_form.tt is where the client side JavaScript lives
that handles/honors passtyhrough redirect, which allows browsing to a
secure URL while not logged in, letting the user login, and then
continuing on to that original URL after successful authentication, has
been modified to use plain JavaScript and no longer relies on ExtJS
calls. This allows this functionality to work even if ExtJS is not
loaded, which we are now supporting (added for Rapi::Blog improvements)

Also removed the JavaScript autofocus logic to focus the username field,
instead the HTML 'autofocus' attribute is set on the input tag itself,
since this feature has now been around for a very long time and it is
more than safe to rely on browser support.

Note: the default full login page still loads the full RapidApp/ExtJS
JavaScript and CSS environment even though it is not actually needed for
the login page. The reason this is done remains a performance issue, to
pre-cache these files so that it doesn't need to happen during the
actual login process. This is still the desired behavior, however, we
don't want to force other uses to do this if they do not need to, such
as optionally in certain Rapi::Blog default pages.

(set dev version 1.3102_02)