SUPPORT NEW RapidApp CONFIG PARAM 'error_template'
New config $c->config->{RapidApp}{error_template} can now be set to
override exception rendering to use the specified template. Rendered
with teh Template Controller with the exception/error in the tt varibale
named 'error'

This is an optional param, and the default is still unset which still uses
the current behaviour which is to defer to the built-in Catalyst::Engine
error code.

Set dev version 1.3102_03