CoreSchema: USERNAME/PASSWORD CHANGES PROPERLY HANDLE ACTIVE SESSIONS
When a user username is changed, the system now considers all active
sessions associated with that user, and if there are any, it munges the
session data to match the new username. This means that not only can a
user change their own username without having to log out and back in,
any of their other active sessions will seamlessly continue to work, as
well as the case of an admin changing another user's username while they
are logged in. Their session will continue to work just fine, and there
will be no impact until they need to login again (they'll obviously need
to know their new username).

Additionally, when a password is changed, the system will instantly
terminate all of that user's active sessions, EXCEPT the session of the
user actually making the chnage, when they are changing their own
password.

(set dev version 1.3108_03)